Extract MFM types
