+++ src/mfm/normalize.ts
@@ -1,6 +1,6 @@
 import * as A from '../prelude/array';
 import * as S from '../prelude/string';
-import { MfmForest, MfmTree } from './parser';
+import { MfmForest, MfmTree } from './types';
 import { createTree, createLeaf } from '../prelude/tree';
 
 function isEmptyTextTree(t: MfmTree): boolean {
