+++ src/misc/extract-mentions.ts
@@ -1,6 +1,6 @@
 // test is located in test/extract-mentions
 
-import { MentionNode, MfmForest } from '../mfm/parser';
+import { MentionNode, MfmForest } from '../mfm/types';
 import { preorderF } from '../prelude/tree';
 
 export default function(mfmForest: MfmForest): MentionNode['props'][] {
